[corlib] Import System.IO.File from CoreFX
